% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputDF.R
\name{createInputDF}
\alias{createInputDF}
\title{Create a combined input data frame.}
\usage{
createInputDF(
  country,
  historic_include = T,
  projection_include = F,
  historicDF = NULL,
  projectionDF = NULL,
  vaccination_scenario = "baseline",
  start_year = 1990,
  end_year = NULL
)
}
\arguments{
\item{country}{The three-digit ISO3 code for the country in which to
estimate measles CFRs. See Details for how to set this up as a loop for
multiple countries.}

\item{historic_include}{Boolean value (TRUE/FALSE). If set to TRUE,
will include default historic  Default set to TRUE.}

\item{projection_include}{Boolean value (TRUE/FALSE). Default set to TRUE.}

\item{historicDF}{Optional parameter. Output data frame created by
\code{createHistoricDF()}.}

\item{projectionDF}{Optional parameter. Output data frame created by
\code{createProjectionDF()}.}

\item{vaccination_scenario}{An indicator to specify whether to estimate
CFRs given vaccination (i.e. baseline) or no-vaccination scenario
(options: baseline, no_vaccination). Default set to baseline.}

\item{start_year}{The start year to begin measles CFR estimation. An
integer class value that is bounded between 1980 and 2100. start_year
cannot be greater than end_year. Default set to 1990.}

\item{end_year}{The end year through which to estimate measles CFR. An
integer class value that is bounded between 1980 and 2100. end_year
cannot be less than start_year.}
}
\value{
A data frame of numeric class for the user-specified start to end
years. The data frame is long by year and wide by covariates.
}
\description{
If historic and/or projection data frames are created by the
user, the function will combine user-specified values with default values
for the specified time period. If no user-specified data frames are created,
the function will rely on default values for the specified time period.
}
\details{
The function returns the default historic (if historic_include is
set to TRUE) and/or projection (if projection_include is set to TRUE) data
frames as a combined data frame. If the user supplies an historic data
frame created by \code{createHistoricDF()} and/or projection data frame
created by \code{createProjectionDF()}, the function returns the combined
data frame of the user-specified data frame objects.
}
\examples{
# Create input DF with package defaults for years 2000 to 2030
eth_inputDF_default <- createInputDF(country="ETH", start_year=2000, end_year=2030)
# Create input DF with user-created historic DF and projection DF
eth_inputDF <- createInputDF(country="ETH", historicDF=eth_historicDF, projectionDF=eth_projectionDF, historic_include=FALSE, projection_include=FALSE, vaccination_scenario="baseline", start_year=1980, end_year=2100)

}
