% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectionDF.R
\name{createProjectionDF}
\alias{createProjectionDF}
\title{Create a projection data frame.}
\usage{
createProjectionDF(
  country,
  df_incidence = NULL,
  df_mcv1 = NULL,
  df_maternal_education = NULL,
  df_u5mr = NULL,
  df_prop_urban = NULL,
  df_vitA = NULL,
  bound_sample_range = F
)
}
\arguments{
\item{country}{The three-digit ISO3 code for the country in which to
estimate measles CFRs. See Details for how to set this up as a loop for
multiple countries.}

\item{df_incidence}{Optional parameter. A country-specific numeric vector of
measles incidence that is long by year. Values must be non-negative.}

\item{df_mcv1}{Optional parameter. A country-specific numeric vector of
MCV1 coverage that is long by year. Values are bounded between 0 and 1.}

\item{df_maternal_education}{Optional parameter. A country-specific numeric
vector of maternal educational attainment that is long by year. Values must
be bounded between 0 and 18 years.}

\item{df_u5mr}{Optional parameter. A country-specific numeric vector of
under-five mortality rate (per 1000) that is long by year. Values must be
non-negative.}

\item{df_prop_urban}{Optional parameter. A country-specific numeric vector
of proportion of population living in urban settings that is long by year.
Values must be between 0 and 1.}

\item{df_vitA}{Optional parameter. A country-specific numeric vector of
vitamin A deficiency prevalence that is long by year. Values must be
between 0 and 1.}

\item{bound_sample_range}{Boolean value (TRUE/FALSE). If set to FALSE
(set as default), the package will provide measles CFR estimates with an
accompanying warning. The sample bounds will be applied to any
visualizations that display covariates in other package functions.
If TRUE, the package will restrict user-specified covariate values to the
sample range that the model was fit to. Please see below (Details section)
for more information.}
}
\value{
A data frame of numeric class for the projection (i.e., future)
period (projectionDF). The data frame is long by year and wide by covariates.
}
\description{
User provides inputs of measles incidence, measles-containing
vaccine first-dose (MCV1) coverage, or other covariates for the projection
period. The projection period can be specified for any end year through the
year 2100. The function will standardize the provided covariates, rely on
default values for any covariates that are not user-specified, and return a
data frame that is long by year and wide by covariates.
}
\details{
First, this function transforms each covariate according to the
defined model fit in Sbarra, et al. (https://doi.org/10.1016/S2214-109X(23)00043-8).

Second, this function standardizes the transformed covariates by subtracting
the mean of the covariate and dividing by the standard deviation. Third, the
function returns a data frame that is long by year and wide by transformed,
standardized covariates.

Each covariate is held constant at its 2019 estimated value across the
projected time period. Across all countries, the sample range of each
untransformed, unstandardized covariate value is as described below.\tabular{lrr}{
   \strong{Covariate} \tab \strong{Minimum Value} \tab \strong{Maximum Value} \cr
   Measles incidence \tab 0.0000 \tab 0.0549 \cr
   Maternal education \tab 1.5164 \tab 13.9316 \cr
   MCV1 coverage \tab 0.37 \tab 0.99 \cr
   Under-5 mortality rate (per 1000) \tab 3.1 \tab 118.3 \cr
   Proportion living in urban setting \tab 0.1325 \tab 0.9120 \cr
   Vitamin A deficiency prevalence \tab 0.0042 \tab 0.2644 \cr
}
}
\examples{
# Set local working directory; revise to location of input files
setwd("E:/")
# Load and save relevant input files to working environment
eth_incidence_proj <- read.csv("eth_incidence.csv",header=TRUE)
eth_mcv1_proj <- read.csv("eth_mcv1.csv",header=TRUE)
# Create projection DF with user-specified incidence and MCV1 coverage
# All other covariates rely on package defaults
eth_projectionDF <- createProjectionDF(country="ETH", df_incidence=eth_incidence_proj, df_mcv1=eth_mcv1_proj)

}
