% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayAgeCFR.R
\name{displayAgeCFR}
\alias{displayAgeCFR}
\title{Display plots of measles case fatality rate by age.}
\usage{
displayAgeCFR(predict_df, list_of_years = c(2000, 2010, 2019))
}
\arguments{
\item{predict_df}{Data frame object created by \code{predictCFR()} function.}

\item{list_of_years}{A list of years obtained with unique values from
\code{predictCFR()} function, or a list specified by the user in order to
display a subset of years.}
}
\value{
A multi-panel plot of estimated measles CFR by single year of age.
The plots are graphed over time (i.e., by calendar year) for the specified
time period.
}
\description{
In a multi-panel plot, each single year of age estimated with
predictDF is visually displayed by calendar year.
}
\details{
The function returns a multi-panel plot in which each year of age
from start_age to end_age included in the \code{predictCFR()} object is
graphed over calendar year. If no year_list is supplied, the function will
return plots with start_year and end_year defined according to the
\code{predictCFR()} object; otherwise, a list of years can be passed to the
function in order to display the relevant calendar years. If
plot_uncertainty is set to “TRUE” (default), the multi-panel plots will
incorporate uncertainty intervals based on the defined upper_bound and
lower_bound percentiles in the \code{predictCFR()} object. If plot_uncertainty
is set to “FALSE,” the multi-panel plots will only include estimated mean
CFRs over time.
}
\examples{
# Examine predicted CFRs using predictions based on default covariates
# Rely on list of years generated through unique values in dataframe
displayAgeCFR(predictDF=eth_cfr_default)
# Examine predicted CFRs for years 2000 to 2030
displayAgeCFR(predictDF=eth_cfr, list_of_years=c(seq(2000,2030,1))

}
