% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayCovariates.R
\name{displayCovariates}
\alias{displayCovariates}
\title{Visually display covariates in individual plots.}
\usage{
displayCovariates(country, inputDF = NULL)
}
\arguments{
\item{country}{The three-digit ISO3 code for the country in which to
estimate measles CFRs. See Details for how to set this up as a loop for
multiple countries.}

\item{inputDF}{Optional parameter. If supplied, the function will display
the combination of user-specified and default covariate values by year. Each
covariate will be displayed in its own plot (i.e., six plots). If no inputDF
is supplied, the function will rely on default covariate values.}
}
\value{
A multi-panel plot of each set of covariate values. The plots are
graphed over time (i.e., by calendar year).
}
\description{
In a multi-panel plot object, each set of covariate values
is visually displayed over time (i.e., by calendar year) from the start
year to end year specified. Covariate values can be user-specified through
the use of the \code{createHistoricDF()}, \code{createProjectionDF()}, and
\code{createInputDF()} functions, or can rely on default covariate values.
}
\details{
The function returns the default covariate values over time
(in calendar years) as a multi-panel plot object. If the user supplies an
input data frame created by \code{createInputDF()}, the function returns the
user-specified covariate values over time (in calendar years) as a
multi-panel plot object.
}
