% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayTimeCFR.R
\name{displayTimeCFR}
\alias{displayTimeCFR}
\title{Display plots of measles case fatality rate over time (i.e., by calendar
year).}
\usage{
displayTimeCFR(predict_df, list_of_ages = c(0, 1, 2, 3, 4))
}
\arguments{
\item{predict_df}{Data frame object created by \code{predictCFR()} function.}

\item{list_of_ages}{A list of single-year ages obtained with unique values
from \code{predictCFR()} function, or a list specified by the user in order
to display a subset of ages}
}
\value{
A multi-panel plot of estimated measles CFR by calendar year. The
plots are graphed by single year of age for the specified ages.
}
\description{
In a multi-panel plot, each calendar estimated with \code{predictCFR()}
is visually displayed by single year of age.
}
\details{
The function returns a multi-panel plot in which each calendar
year from start_year to end_year included in the \code{predictCFR()} object
is graphed for single-year age groups. If no age_list is supplied, the
function will return plots with start_age and end_age defined according
to the \code{predictCFR()} object; otherwise, a list of single-year ages
can be passed to the function in order to display the relevant age groups.
If plot_uncertainty is set to “TRUE” (default), the multi-panel plots will
incorporate uncertainty intervals based on the defined upper_bound and
lower_bound percentiles in the \code{predictCFR()} object. If plot_uncertainty
is set to “FALSE,” the multi-panel plots will only include estimated mean
CFRs by age.
}
\examples{
# Examine predicted CFRs using predictions based on default covariates
# Rely on list of ages generated through unique values in dataframe
displayTimeCFR(predict_df=eth_cfr_default)
# Examine predicted CFRs for ages 0 to 14
displayAgeCFR(predict_df=eth_cfr, list_of_ages=c(seq(0,14,1))

}
