% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOfCountries.R
\name{listOfCountries}
\alias{listOfCountries}
\title{List all ISO3 countries that are accepted.}
\usage{
listOfCountries()
}
\value{
ISO3 countries
}
\description{
A list of of all low- and middle-income ISO3 countries that are accepted by the measlesCFR package.
}
